+++ src/index.ts
@@ -1,0 +1,10 @@
+// 导出所有服务和类型
+export * from "./services/configService";
+export * from "./services/contextAnalyzer";
+export * from "./services/styleDetector";
+export * from "./services/llmService";
+export * from "./ui/statusBarManager";
+export * from "./commands/commandManager";
+export * from "./types/context";
+export * from "./enums/configKey";
+export * from "./enums/varStyle";